First parsing out of classes
