+++ SimPEG/directives/maps.py
@@ -1,0 +1,51 @@
+from ..utils import (
+    spherical2cartesian,
+    cartesian2spherical,
+)
+from .base import InversionDirective
+
+
+class ProjectSphericalBounds(InversionDirective):
+    r"""
+    Trick for spherical coordinate system.
+    Project :math:`\theta` and :math:`\phi` angles back to :math:`[-\pi,\pi]`
+    using back and forth conversion.
+    spherical->cartesian->spherical
+    """
+
+    def initialize(self):
+        x = self.invProb.model
+        # Convert to cartesian than back to avoid over rotation
+        nC = int(len(x) / 3)
+
+        xyz = spherical2cartesian(x.reshape((nC, 3), order="F"))
+        m = cartesian2spherical(xyz.reshape((nC, 3), order="F"))
+
+        self.invProb.model = m
+
+        for sim in self.simulation:
+            sim.model = m
+
+        self.opt.xc = m
+
+    def endIter(self):
+        x = self.invProb.model
+        nC = int(len(x) / 3)
+
+        # Convert to cartesian than back to avoid over rotation
+        xyz = spherical2cartesian(x.reshape((nC, 3), order="F"))
+        m = cartesian2spherical(xyz.reshape((nC, 3), order="F"))
+
+        self.invProb.model = m
+
+        phi_m_last = []
+        for reg in self.reg.objfcts:
+            reg.model = self.invProb.model
+            phi_m_last += [reg(self.invProb.model)]
+
+        self.invProb.phi_m_last = phi_m_last
+
+        for sim in self.simulation:
+            sim.model = m
+
+        self.opt.xc = m
